+++ ecomel-api/src/main/java/br/com/ecomel/service/UsuarioService.java
@@ -66,7 +66,7 @@ public class UsuarioService {
         carteira.setUsuario(usuario);
         carteira.setCodigoEndereco(gerarNovoCodigoValido());
         carteira.setAtivo(true);
-        carteira.setTokenEcomel(BigInteger.ZERO);
+        carteira.setTokenEcomel(BigDecimal.ZERO);
         carteira.setSaldoFavos(BigDecimal.ZERO);
         carteira.setCriadoEm(audit.data());
         carteira.setCriadoPor(audit.usuario());
@@ -106,7 +106,7 @@ public class UsuarioService {
         Usuario usuario = usuarioRepository.findById(id)
                 .orElseThrow(() -> new BusinessException("Usuário não encontrado"));
         
-        if (usuario.getCarteira().getTokenEcomel().compareTo(BigInteger.ZERO) > 0 || 
+        if (usuario.getCarteira().getTokenEcomel().compareTo(BigDecimal.ZERO) > 0 || 
             usuario.getCarteira().getSaldoFavos().compareTo(BigDecimal.ZERO) > 0) {
             throw new BusinessException("Conta possui ativos. Zere os saldos antes de desativar.");
         }
